Document secret params
